First implementation JUnit Test

+++ src/main/java/com/eStore/bear/product/service/ProductService.java
@@ -4,13 +4,16 @@ import com.eStore.bear.product.dto.Product;
 import com.eStore.bear.product.exception.CurrencyValidException;
 import com.eStore.bear.product.exception.ProductValidException;
 import com.eStore.bear.product.repository.ProductRepository;
+import com.eStore.bear.product.response.ProductResponse;
 import com.eStore.bear.product.service.config.ProductConfiguration;
 import com.eStore.bear.product.utils.ProductConstants;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -22,44 +25,61 @@ public class ProductService {
     @Autowired
     private ProductConfiguration productConfiguration;
 
-    public ProductService(ProductRepository productRepository) {
+    public ProductService(ProductRepository productRepository, ProductConfiguration productConfiguration) {
         this.productRepository = productRepository;
+        this.productConfiguration = productConfiguration;
     }
 
-    public String addProduct(Product product) {
+    public ProductResponse addProduct(Product product) {
 
-        if(product.getPrice() == 0 && product.getDiscount() > 0) {
+        if (product.getPrice() == 0 && product.getDiscount() > 0) {
             throw new ProductValidException(ProductConstants.PRODUCT_DISCOUNT);
         }
 
-        if(!productConfiguration.getCurrencies().contains(product.getCurrency())) {
+        if (!productConfiguration.getCurrencies().contains(product.getCurrency())) {
             throw new CurrencyValidException(ProductConstants.PRODUCT_CURRENCY + productConfiguration.getCurrencies());
         }
 
         productRepository.save(product);
-        return "product added successfully.";
+        return new ProductResponse(HttpStatus.OK, product.getName() + " added successfully in the Store");
     }
 
     public List<Product> listAllProducts() {
-        return productRepository.findAll();
+        List<Product> products = productRepository.findAll();
+        if (products == null || products.isEmpty()) {
+            throw new ProductValidException("Products not found in the Store");
+        }
+        return products;
     }
 
     public List<Product> productCategoryList(String category) {
+        List<Product> productsByCategory = productRepository.findByCategory(category);
 
-        return productRepository.findByCategory(category);
+        if(productsByCategory == null || productsByCategory.isEmpty()) {
+            throw new ProductValidException("Products not found for the category" + category + " in the Store");
+        }
+        return productsByCategory;
     }
 
     public Product productById(String id) {
-        return productRepository.findById(id).get();
+        return productRepository.findById(id).orElseThrow(() -> new ProductValidException("Product not found for id - " + id));
     }
 
-    public String updateProduct(Product product) {
+    public ProductResponse updateProduct(Product product) {
+        Optional<Product> prd =  productRepository.findById(product.getId());
+        if(!prd.isPresent()) {
+            return new ProductResponse(HttpStatus.NOT_FOUND, "Product to be updated isn't found to the Store.");
+        }
         productRepository.save(product);
-        return "product updated successfully";
+        return new ProductResponse(HttpStatus.OK, "Product updated successfully.");
     }
 
-    public String deleteProductById(String id) {
+    public ProductResponse deleteProductById(String id) {
+        Optional<Product> product =  productRepository.findById(id);
+        if(!product.isPresent()) {
+            return new ProductResponse(HttpStatus.NOT_FOUND, "Product to be deleted isn't found to the Store.");
+        }
         productRepository.deleteById(id);
-        return "product delected successfully.";
+        return new ProductResponse(HttpStatus.OK, "Product deleted successfully.");
     }
 }
